+++ tencentcloud/service_tencentcloud_scf.go
@@ -2,6 +2,7 @@ package tencentcloud
 
 import (
 	"context"
+	"strings"
 
 	"github.com/hashicorp/terraform-plugin-sdk/helper/resource"
 	"github.com/pkg/errors"
@@ -89,14 +90,18 @@ func (me *ScfService) CreateFunction(ctx context.Context, info scfFunctionInfo) 
 		ratelimit.Check(request.GetAction())
 
 		if _, err := client.CreateFunction(request); err != nil {
-			return retryError(errors.WithStack(err), "InternalError")
+			e, ok := err.(*sdkErrors.TencentCloudSDKError)
+			if ok && strings.Contains(e.Code, "ResourceInUse") {
+				return resource.NonRetryableError(err)
+			}
+			return retryError(errors.WithStack(err))
 		}
 		return nil
 	}); err != nil {
 		return err
 	}
 
-	return waitScfFunctionReady(ctx, info.name, *info.namespace, client)
+	return nil
 }
 
 func (me *ScfService) DescribeFunction(ctx context.Context, name, namespace string) (resp *scf.GetFunctionResponse, err error) {